#OSIS-1203 Modifications des tests
