+++ base/templatetags/learning_unit.py
@@ -121,14 +121,13 @@ def get_style_of_label_text(label_text, style, title):
 
 @register.filter
 def get_previous_acronym(luy):
-    if has_proposal:
+    if has_proposal(luy):
         proposal = ProposalLearningUnit.objects \
             .filter(learning_unit_year=luy) \
             .order_by('-learning_unit_year__academic_year__year').first()
-        if proposal:
-            return proposal.initial_data.get('learning_unit_year').get('acronym', None)
+        if proposal and proposal.initial_data and proposal.initial_data.get('learning_unit_year'):
+            return proposal.initial_data.get('learning_unit_year', None)
     else:
-        luy = find_lt_learning_unit_year_with_different_acronym(self)
-        if luy:
-            return luy.acronym
+        luy = find_lt_learning_unit_year_with_different_acronym(luy)
+        return luy.acronym if luy else None
     return None
